+++ gradle.properties
@@ -8,5 +8,5 @@ plugin_group=Uggles
 plugin_description=Adds a combat overhaul to Hytale
 plugin_author=Uggles
 plugin_website=
-plugin_main_entrypoint=com.example.exampleplugin.ExamplePlugin
+plugin_main_entrypoint=com.abilities.abilitiesplugin.CombatAbilityPlugin
 server_version=*
