+++ src/main/java/com/abilities/abilitiesplugin/.vs/exampleplugin/FileContentIndex/0d71dd57-d2c8-40fe-b222-045d4b979ad6.vsidx
