+++ src/main/java/com/abilities/abilitiesplugin/AbilityHotbarState.java
@@ -1,4 +1,4 @@
-package com.example.exampleplugin;
+package com.abilities.abilitiesplugin;
 
 import java.util.HashMap;
 import java.util.Map;
