+++ src/main/java/com/abilities/abilitiesplugin/AbilityRegistry.java
@@ -1,4 +1,4 @@
-package com.example.exampleplugin;
+package com.abilities.abilitiesplugin;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
