+++ src/main/java/com/abilities/abilitiesplugin/IAbilityPlugin.java
@@ -1,4 +1,4 @@
-package com.example.exampleplugin;
+package com.abilities.abilitiesplugin;
 
 public interface IAbilityPlugin {
     boolean CAO_DoAbility(PackagedAbilityData Data, AbilityContext Context);
