+++ src/main/java/com/abilities/abilitiesplugin/PackagedAbilityData.java
@@ -1,4 +1,4 @@
-package com.example.exampleplugin;
+package com.abilities.abilitiesplugin;
 
 public final class PackagedAbilityData {
