+++ src/main/java/com/abilities/abilitiesplugin/WeaponDefinition.java
@@ -1,4 +1,4 @@
-package com.example.exampleplugin;
+package com.abilities.abilitiesplugin;
 
 import java.util.List;
