+++ src/main/java/com/abilities/abilitiesplugin/WeaponRegistry.java
@@ -1,4 +1,4 @@
-package com.example.exampleplugin;
+package com.abilities.abilitiesplugin;
 
 import com.google.gson.*;
 import java.io.InputStream;
