--- src/main/java/com/example/exampleplugin/ExampleCommand.java
@@ -1,28 +1,0 @@
-package com.example.exampleplugin;
-
-import com.hypixel.hytale.protocol.GameMode;
-import com.hypixel.hytale.server.core.Message;
-import com.hypixel.hytale.server.core.command.system.CommandContext;
-import com.hypixel.hytale.server.core.command.system.basecommands.CommandBase;
-
-import javax.annotation.Nonnull;
-
-/**
- * This is an example command that will simply print the name of the plugin in chat when used.
- */
-public class ExampleCommand extends CommandBase {
-    private final String pluginName;
-    private final String pluginVersion;
-
-    public ExampleCommand(String pluginName, String pluginVersion) {
-        super("test", "Prints a test message from the " + pluginName + " plugin.");
-        this.setPermissionGroup(GameMode.Adventure); // Allows the command to be used by anyone, not just OP
-        this.pluginName = pluginName;
-        this.pluginVersion = pluginVersion;
-    }
-
-    @Override
-    protected void executeSync(@Nonnull CommandContext ctx) {
-        ctx.sendMessage(Message.raw("Hello from the " + pluginName + " v" + pluginVersion + " plugin!"));
-    }
-}
