--- src/main/resources/Server/Item/Items/Example_Block.json
@@ -1,56 +1,0 @@
-{
-  "TranslationProperties": {
-    "Name": "server.items.Example_Block.name"
-  },
-  "ItemLevel": 600,
-  "Icon": "Icons/ItemsGenerated/Example_Block.png",
-  "Categories": [
-    "Blocks.Soils"
-  ],
-  "PlayerAnimationsId": "Block",
-  "Set": "Soil",
-  "BlockType": {
-    "Material": "Solid",
-    "DrawType": "Cube",
-    "Group": "Dirt",
-    "Flags": {},
-    "Gathering": {
-      "Breaking": {
-        "GatherType": "Soils"
-      }
-    },
-    "BlockParticleSetId": "Sand",
-    "Textures": [
-      {
-        "All": "BlockTextures/Example_Block.png",
-        "Weight": 1
-      }
-    ],
-    "ParticleColor": "#3d3e3e",
-    "BlockSoundSetId": "Gravel",
-    "TransitionTexture": "BlockTextures/Transition_Ashes.png",
-    "TransitionToGroups": [
-      "Stone",
-      "Lava",
-      "Dirt"
-    ]
-  },
-  "Tags": {
-    "Type": [
-      "Soil"
-    ]
-  },
-  "ItemSoundSetId": "ISS_Blocks_Soft",
-  "IconProperties": {
-    "Scale": 0.58823,
-    "Rotation": [
-      22.5,
-      45,
-      22.5
-    ],
-    "Translation": [
-      0,
-      -13.5
-    ]
-  }
-}
