--- src/main/resources/Server/Item/RootInteractions/Ability_Test.json
@@ -1,5 +1,0 @@
-{
-  "Interactions": [
-    "Explode_Generic_Blocks"
-  ]
-}
