--- src/main/resources/Server/Item/RootInteractions/SomeRoot.json
@@ -1,1 +1,0 @@
-{}
